light is the problem, updated readme and all
