+++ src/pages/certifications/index.tsx
@@ -3,13 +3,26 @@ import { NextPage } from 'next';
 import Layout from '@/components/Layout';
 import { Container, Title } from '@/styles/global.styles';
 import Certification from '@/components/Certification';
+import certifications from '../../assets/data/certifications.json';
+import tw, { styled } from 'twin.macro';
+
+const CertificationContainer = styled.div`
+  ${tw`
+  flex
+  flex-row
+`}
+`;
 
 const CertificationsPage: NextPage = () => {
   return (
     <Layout title="Certifications">
       <Container>
         <Title>These are my Certifications and Skills</Title>
-        <Certification />
+        <CertificationContainer>
+          {certifications.map((certification) => {
+            return <Certification data={certification} />;
+          })}
+        </CertificationContainer>
       </Container>
     </Layout>
   );
